+++ typings/index.d.ts
@@ -4,7 +4,9 @@ declare module '@thumbtack/thumbprint-tokens';
 declare module 'intersection-observer';
 declare module 'object-fit-images';
 
+type CSSModuleType = { [key: string]: string };
+
 declare module '*.module.scss' {
-    const classes: any;
+    const classes: CSSModuleType;
     export default classes;
 }