+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Changed
+
+-   [Patch] Rewrite `FormNote` component in TypeScript.
+-   [Patch] Rewrite `Grid` and `GridColumn` components in TypeScript.
+-   [Patch] Rewrite `LoaderDots` component in TypeScript.
+
 ## 9.5.0 - 2019-09-24
 
 ### Added
@@ -10,7 +16,7 @@
 
 ## 9.4.1 - 2019-09-24
 
-### Added
+### Changed
 
 -   [Patch] Rewrite `ButtonRow` component in TypeScript.